+++ compiler-cli/src/fs.rs
@@ -12,7 +12,7 @@ use gleam_core::{
 };
 use gleam_language_server::{DownloadDependencies, Locker, MakeLocker};
 use std::{
-    collections::HashSet,
+    collections::{HashSet, VecDeque},
     fmt::Debug,
     fs::{File, exists},
     io::{self, BufRead, BufReader, Write},
@@ -179,6 +179,10 @@ impl FileSystemWriter for ProjectIO {
 
     fn hardlink(&self, from: &Utf8Path, to: &Utf8Path) -> Result<(), Error> {
         hardlink(from, to)
+    }
+
+    fn hardlink_directory(&self, from: &Utf8Path, to: &Utf8Path) -> Result<(), Error> {
+        hardlink_directory(from, to)
     }
 
     fn symlink_dir(&self, from: &Utf8Path, to: &Utf8Path) -> Result<(), Error> {
@@ -662,6 +666,11 @@ pub fn copy_dir(
     .map(|_| ())
 }
 
+/// Symlink directory.
+/// If it couldn't symlink directory on Windows because of error 1314, which
+/// occurs without Developer Mode enabled, it falls back to hardlinking each
+/// file in directory.
+///
 pub fn symlink_dir(
     src: impl AsRef<Utf8Path> + Debug,
     dest: impl AsRef<Utf8Path> + Debug,
@@ -670,17 +679,25 @@ pub fn symlink_dir(
     let src = canonicalise(src.as_ref())?;
 
     #[cfg(target_family = "windows")]
-    let result = std::os::windows::fs::symlink_dir(src, dest.as_ref());
+    let result = std::os::windows::fs::symlink_dir(&src, dest.as_ref());
     #[cfg(not(target_family = "windows"))]
-    let result = std::os::unix::fs::symlink(src, dest.as_ref());
+    let result = std::os::unix::fs::symlink(&src, dest.as_ref());
 
-    result.map_err(|err| Error::FileIo {
-        action: FileIoAction::Link,
-        kind: FileKind::File,
-        path: Utf8PathBuf::from(dest.as_ref()),
-        err: Some(err.to_string()),
-    })?;
-    Ok(())
+    match result {
+        Ok(()) => Ok(()),
+
+        // Fallback to hardlinking if we can't symlink. This occurs on Windows
+        // without Developer Mode enabled. We match on the raw OS code, since
+        // this error has no specific error kind.
+        #[cfg(target_family = "windows")]
+        Err(err) if err.raw_os_error() == Some(1314) => hardlink_directory(src, dest),
+        Err(err) => Err(Error::FileIo {
+            action: FileIoAction::Link,
+            kind: FileKind::File,
+            path: Utf8PathBuf::from(dest.as_ref()),
+            err: Some(err.to_string()),
+        }),
+    }
 }
 
 pub fn hardlink(
@@ -696,6 +713,55 @@ pub fn hardlink(
             err: Some(err.to_string()),
         })
         .map(|_| ())
+}
+
+/// Hardlink directory.
+/// This is done by deleting destination directory if it exists and recursively
+/// creating directories and hardlinking files inside the directory.
+///
+pub fn hardlink_directory(
+    src: impl AsRef<Utf8Path> + Debug,
+    dest: impl AsRef<Utf8Path> + Debug,
+) -> Result<(), Error> {
+    tracing::trace!(src=?src, dest=?dest, "hardlinking_dir");
+
+    let src = src.as_ref();
+    let dest = dest.as_ref();
+
+    // Recreate destination directory to avoid presence of old files,
+    // that don't exist in source directory already.
+    delete_directory(dest)?;
+
+    let mut stack: VecDeque<(Utf8PathBuf, Utf8PathBuf)> = VecDeque::new();
+    stack.push_back((src.to_path_buf(), dest.to_path_buf()));
+
+    while let Some((source_directory, destination_directory)) = stack.pop_front() {
+        mkdir(&destination_directory)?;
+        for entry in read_dir(source_directory)? {
+            let entry = entry.map_err(|e| Error::FileIo {
+                action: FileIoAction::Read,
+                kind: FileKind::Directory,
+                path: src.to_path_buf(),
+                err: Some(e.to_string()),
+            })?;
+
+            let path = entry.path().to_path_buf();
+
+            let file_name = path.file_name().ok_or_else(|| Error::NonUtf8Path {
+                path: path.to_path_buf().into(),
+            })?;
+
+            let destination = destination_directory.join(file_name).to_path_buf();
+
+            if path.is_dir() {
+                stack.push_back((path, destination));
+            } else {
+                hardlink(path, destination)?;
+            }
+        }
+    }
+
+    Ok(())
 }
 
 /// Check if the given path is inside a git work tree.