+++ compiler-core/src/io/memory.rs
@@ -189,6 +189,10 @@ impl FileSystemWriter for InMemoryFileSystem {
         Ok(())
     }
 
+    fn hardlink_directory(&self, _: &Utf8Path, _: &Utf8Path) -> Result<(), Error> {
+        panic!("unimplemented") // TODO
+    }
+
     fn symlink_dir(&self, _: &Utf8Path, _: &Utf8Path) -> Result<(), Error> {
         panic!("unimplemented") // TODO
     }