+++ language-server/src/tests.rs
@@ -210,6 +210,10 @@ impl FileSystemWriter for LanguageServerTestIO {
         self.io.hardlink(from, to)
     }
 
+    fn hardlink_directory(&self, from: &Utf8Path, to: &Utf8Path) -> Result<()> {
+        self.io.hardlink_directory(from, to)
+    }
+
     fn symlink_dir(&self, from: &Utf8Path, to: &Utf8Path) -> Result<()> {
         self.io.symlink_dir(from, to)
     }